+++ src/termout.c
@@ -2075,7 +2075,6 @@ do_csi(uchar c)
   if (arg1 < 0)
     arg1 = 0;
   int arg0_def1 = arg0 ?: 1;  // first arg with default 1
-  static int ime_status = 0;
 
   // DECRQM quirk
   if (term.esc_mod == 0xFF && esc_mod0 == '?' && esc_mod1 == '$' && c == 'p')
@@ -2591,15 +2590,12 @@ do_csi(uchar c)
         insdel_column(curs->x, true, arg0_def1);
     when CPAIR('#', 't'):  /* application scrollbar */
       win_set_scrollview(arg0, arg1);
-    // IME control
-    when CPAIR('<', 'r'):
-      ImmSetOpenStatus(imc, ime_status);
-      win_set_ime_open(ime_status);
-    when CPAIR('<', 's'):
-      ime_status = ImmGetOpenStatus(imc);
-    when CPAIR('<', 't'):
-      ImmSetOpenStatus(imc, arg0);
-      win_set_ime_open(arg0);
+    when CPAIR('<', 't'):  /* TTIMEST: change IME state (Tera Term) */
+      win_set_ime(arg0);
+    when CPAIR('<', 's'):  /* TTIMESV: save IME state (Tera Term) */
+      push_mode(-1, win_get_ime());
+    when CPAIR('<', 'r'):  /* TTIMERS: restore IME state (Tera Term) */
+      win_set_ime(pop_mode(-1));
   }
 }